Updated metadata.json
The license in metadata.json must be present in the SPDX License List
(http://spdx.org/licenses) for puppetforge to detect the correct
license. The name of the Apache license was misspelled in metadata.json,
which had a negetive effect on Metadata Quality score. Also the
metadata.json has to include to supported operating systems.

While this module is probably supported on any system that uses RPM, I
have only tested it on RHEL5 and RHEL6.