(#6) Support puppet:// links as source
In some situations it is desireable to use your puppet master as a
central resource for gpg keys. Since rpm does not support `puppet://`
URLs you'd have to make sure that you install the key with a seperate
`file` resource.

To solve this, check the `source` before runnig `rpm --import`:

* if source is an absolute path, use it directly for `rpm --import`
* if source is a file:// resource, convert the source to a local path
  and use this for `rpm --import`
* if source is a `puppet://` URL and puppet runs locally (via puppet
  apply), get the local path to the file and use that `rpm --import`
* if source is a `puppet://` URL and puppet runs as an agent,
  write the remote content to a temporary file and use the path to the
  temporary file for `rpm --import`
* if source is anything else (e.g. a `http://` link) pass it directly
  to `rpm --import` and hope `rpm` can handle it directly.